+++ packages/core/src/component.ts
@@ -3,9 +3,16 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-import {Constructor, Provider, BoundValue, Binding} from '@loopback/context';
-import {Server} from './server';
+import {
+  Binding,
+  BoundValue,
+  Constructor,
+  createBindingFromClass,
+  Provider,
+} from '@loopback/context';
 import {Application, ControllerClass} from './application';
+import {LifeCycleObserver} from './lifecycle';
+import {Server} from './server';
 
 /**
  * A map of provider classes to be bound to a context
@@ -61,6 +68,8 @@ export interface Component {
     [name: string]: Constructor<Server>;
   };
 
+  lifeCycleObservers?: Constructor<LifeCycleObserver>[];
+
   /**
    * An array of bindings to be aded to the application context. For example,
    * ```ts
@@ -87,13 +96,19 @@ export interface Component {
 export function mountComponent(app: Application, component: Component) {
   if (component.classes) {
     for (const classKey in component.classes) {
-      app.bind(classKey).toClass(component.classes[classKey]);
+      const binding = createBindingFromClass(component.classes[classKey], {
+        key: classKey,
+      });
+      app.add(binding);
     }
   }
 
   if (component.providers) {
     for (const providerKey in component.providers) {
-      app.bind(providerKey).toProvider(component.providers[providerKey]);
+      const binding = createBindingFromClass(component.providers[providerKey], {
+        key: providerKey,
+      });
+      app.add(binding);
     }
   }
 
@@ -112,6 +127,12 @@ export function mountComponent(app: Application, component: Component) {
   if (component.servers) {
     for (const serverKey in component.servers) {
       app.server(component.servers[serverKey], serverKey);
+    }
+  }
+
+  if (component.lifeCycleObservers) {
+    for (const observer of component.lifeCycleObservers) {
+      app.lifeCycleObserver(observer);
     }
   }
 }